+++ changelog.md
@@ -31,6 +31,7 @@
 - Added class `Blog`
 - Added class `Permissions`
 - Added class `RichText`
+- Added class `Reactions`
 - **[BREAKING]** Renamed method and class `GraphExecution` to `GraphExecutions`
 - **[BREAKING]** Renamed method and class `Job` to `Jobs`
 - **[BREAKING]** Renamed method and class `Repository` to `Repositories`